--- a/.gitignore
+++ b/.gitignore
@@ -7,4 +7,5 @@ Pipfile
 Pipfile.lock
 */__pycache__
 *.pyc
+.pytest_cache