--- a/.gitignore
+++ b/.gitignore
@@ -8,4 +8,5 @@ Pipfile.lock
 */__pycache__
 *.pyc
 .pytest_cache
+*.log