… .python-version
… README.md
… pyproject.toml
… tests/test_utils.py
… uv.lock
… visual_excuses/utils.py