**Done**:
- Remove global variables
- Separated into multiple files for clarity
- Separated into multiple functions
- Unit tests
- Refactor some functions
  - Use more f-strings
  - Use get from dict
  - Speed up some loops
- Add better logging

**Notes on performance**:

_visual_excuses.utils.search_teams_

original: 72.598 ± 16.803 µs
new:      29.288 ±  5.100 µs

Approximately 2.5x faster, since this function is called many times for all teams, it speed is noticeable, reducing ms from the run.


_visual_excuses.utils.list_teams_

original: 243.374 ± 30.922 µs 
new: 70.399 ± 19.075 µs

Approximately 3.5x faster, this won't be noticeable, since it is called once, and 243 µs is virtually the same as 70 µs.