@mclemenceau 2025-03-14
Hey @gabrielhsc95,

First of all thank you or much for you interest and the effort you put into into this PR!

There's a lot of changes in that PR and ideally I would like to focus on functional changes and have them in separates PR rather than a big one. I would also recommend to leave the typo and linting aside for now.  Ideally I should have my own linting preference and a github action that enforce it but that is for another day.

I have some general ideas on what I would like to do with visual-excuses in the future that I should probably highlight in a CONTRIBUTING.md or on the README.md. I'll get on that.

As far as performance, goes to me the biggest culprit is how long it can take any web browser to open a busy pyvis visualization and I would like to look into that which may require to use a different visualization toolkit. One of the change I will likely work on next is added abstraction between the excuses data structure and the modeling framework. They are currently tied together which is not great.

So I'll reject the PR for now but look forward for smaller PR in the future with isolated functional improvements.

Thank you
Matthieu